Rename TerminalKeyListener to TerminalViewClient
